new var and logic for it

+++ scripts/Update_ALL.sh
@@ -38,16 +38,28 @@ then
     if [[ ${Enable_Node_Autoupdate} != "true" ]]
     then
         exitVar=1
+    else
+        myNodeAutoupdate=1
     fi
 fi
-shopt -u nocasematch
 
-shopt -s nocasematch
 if [[ ! -z ${Enable_Scripts_Autoupdate} || -n ${Enable_Scripts_Autoupdate} ]]
 then
     if [[ ${Enable_Scripts_Autoupdate} != "true" ]]
     then
         exitVar=2
+    else
+        myScriptsAutoupdate=1
+    fi
+fi
+
+if [[ ! -z ${newReleaseSndMsg} || -n ${newReleaseSndMsg} ]]
+then
+    if [[ ${newReleaseSndMsg} != "true" ]]
+    then
+        exitVar=3
+    else
+      myNewReleaseSndMsg=1
     fi
 fi
 shopt -u nocasematch
@@ -63,6 +75,13 @@ elif [[ $exitVar -eq 2 ]]
 then
     echo "+++INFO: $(basename "$0") Variable Enable_Scripts_Autoupdate not set to true in env.sh"
     echo "+++INFO: Will not update scripts"
+    echo "+++INFO: $(basename "$0") FINISHED $(date +%s) / $(date  +'%F %T %Z')"
+    echo "================================================================================================"
+    exit 1
+elif [[ $exitVar -eq 3 ]]
+then
+    echo "+++INFO: $(basename "$0") Variable newReleaseSndMsg not set to true in env.sh"
+    echo "+++INFO: Will not send notifications abiout new scripts release"
     echo "+++INFO: $(basename "$0") FINISHED $(date +%s) / $(date  +'%F %T %Z')"
     echo "================================================================================================"
     exit 1
@@ -84,53 +103,59 @@ fi
 
 ###############################################################
 #===========================================================
-# Update env.sh for new security update
-if [[ -n "$(cat ${SCRIPT_DIR}/env.sh|grep 'Enable_Autoupdate')" ]];then
-    echo '---WARN: Security update!! Made separate variables for update the node and scripts by the user choice. Setted SCRIPS AUTOUPDATE to FALSE. If you fully trust me, you can enable autoupdate scripts in env.sh by set variable "Enable_Scripts_Autoupdate" to "true"'
-    "${SCRIPT_DIR}/Send_msg_toTelBot.sh" "$HOSTNAME Server" "$Tg_Warn_sign"+'WARN: Security update!! Made separate variables for update the node and scripts by the user choice. Setted SCRIPS AUTOUPDATE to FALSE. If you fully trust me, you can enable autoupdate scripts in env.sh by set variable "Enable_Scripts_Autoupdate" to "true"' 2>&1 > /dev/null
-
-    sed -i.bak 's/Enable_Autoupdate=.*/Enable_Node_Autoupdate=true             # will automatically update rnode, rconsole, tonos-cli etc../' "${SCRIPT_DIR}/env.sh"
-    sed -i.bak '/Enable_Node_Autoupdate/a Enable_Scripts_Autoupdate=false      # Updating scripts. NB! Change it to true if you fully trust me ONLY!!' "${SCRIPT_DIR}/env.sh"
+if [[ "$Custler_Scripts_local_commit" != "$Custler_Scripts_remote_commit" ]]
+then
+    echo '---WARN: Set Enable_Node_Autoupdate to true in env.sh for automatically security updates!! If you fully trust me, you can enable autoupdate scripts in env.sh by set variable "Enable_Scripts_Autoupdate" to "true"'
+    if [[ $myNewReleaseSndMsg -eq 1 ]]
+    then
+        "${SCRIPT_DIR}/Send_msg_toTelBot.sh" "$HOSTNAME Server" "$Tg_Warn_sign"+'WARN: Security info! **NEW** release arrived! But Enable_Node_Autoupdate setted to false and you should upgrade node manually as fast as you can! If you fully trust me, you can enable autoupdate scripts in env.sh by set variable "Enable_Scripts_Autoupdate" to "true"' 2>&1 > /dev/null
+    fi
 fi
+# Update env.sh for new security update
+# ok, here we use force to set autoupdate, strannen'ko =)
+# sed -i.bak 's/Enable_Autoupdate=.*/Enable_Node_Autoupdate=true             # will automatically update rnode, rconsole, tonos-cli etc../' "${SCRIPT_DIR}/env.sh"
+# sed -i.bak '/Enable_Node_Autoupdate/a Enable_Scripts_Autoupdate=false      # Updating scripts. NB! Change it to true if you fully trust me ONLY!!' "${SCRIPT_DIR}/env.sh"
 ###############################################################
 
 #===========================================================
 # Update scripts, if enabled, or send msg re update
-if [[ "$Custler_Scripts_local_commit" == "$Custler_Scripts_remote_commit" ]];then
-    echo "---INFO: Scripts is up to date"
-else
-    if $Enable_Scripts_Autoupdate ;then
-        echo "---INFO: SCRIPTS going to update from $Custler_Scripts_local_commit to new commit $Custler_Scripts_remote_commit"
-        "${SCRIPT_DIR}/Send_msg_toTelBot.sh" "$HOSTNAME Server" "$Tg_Warn_sign INFO: SCRIPTS going to update from $Custler_Scripts_local_commit to new commit $Custler_Scripts_remote_commit" 2>&1 > /dev/null
-
-        Remote_Repo_URL="$(git remote show origin | grep 'Fetch URL' | awk '{print $3}')"
-        echo "---INFO: Update scripts from repo $Remote_Repo_URL"
-
-        #=======================================
-        mkdir -p ${HOME}/Custler_tmp
-        cp -f ${SCRIPT_DIR}/env.sh ${HOME}/Custler_tmp/
-        cp -f ${SCRIPT_DIR}/TlgChat.json ${HOME}/Custler_tmp/
-        cp -f ${SCRIPT_DIR}/RC_Addr_list.json ${HOME}/Custler_tmp/
-
-        git reset --hard
-        git pull --ff-only
-
-        cp -f  ${HOME}/Custler_tmp/env.sh ${SCRIPT_DIR}/
-        cp -f  ${HOME}/Custler_tmp/TlgChat.json ${SCRIPT_DIR}/
-        cp -f  ${HOME}/Custler_tmp/RC_Addr_list.json ${SCRIPT_DIR}/
-        #=======================================
-
-        cat ${SCRIPT_DIR}/Update_Info.txt
-        echo
-        echo "---INFO: SCRIPTS updated. Files env.sh TlgChat.json RC_Addr_list.json keeped."
-        "${SCRIPT_DIR}/Send_msg_toTelBot.sh" "$HOSTNAME Server" "$Tg_CheckMark $(cat ${SCRIPT_DIR}/Update_Info.txt)" 2>&1 > /dev/null
-        "${SCRIPT_DIR}/Send_msg_toTelBot.sh" "$HOSTNAME Server" "$Tg_CheckMark INFO: SCRIPTS updated. Files env.sh TlgChat.json RC_Addr_list.json keeped." 2>&1 > /dev/null
+if [[ ${myScriptsAutoupdate} -eq 1 ]]
+then
+    if [[ "$Custler_Scripts_local_commit" == "$Custler_Scripts_remote_commit" ]];then
+        echo "---INFO: Scripts is up to date"
     else
-        echo '---WARN: Scripts repo was updated. Please check it and update by hand. If you fully trust me, you can enable autoupdate scripts in env.sh by set variable "Enable_Scripts_Autoupdate" to "true"'
-        "${SCRIPT_DIR}/Send_msg_toTelBot.sh" "$HOSTNAME Server" "$Tg_Warn_sign"+'WARN: Scripts repo was updated. Please check it and update. If you fully trust me, you can enable autoupdate scripts in env.sh by set variable "Enable_Scripts_Autoupdate" to "true"' 2>&1 > /dev/null
+        if $Enable_Scripts_Autoupdate ;then
+            echo "---INFO: SCRIPTS going to update from $Custler_Scripts_local_commit to new commit $Custler_Scripts_remote_commit"
+            "${SCRIPT_DIR}/Send_msg_toTelBot.sh" "$HOSTNAME Server" "$Tg_Warn_sign INFO: SCRIPTS going to update from $Custler_Scripts_local_commit to new commit $Custler_Scripts_remote_commit" 2>&1 > /dev/null
+
+            Remote_Repo_URL="$(git remote show origin | grep 'Fetch URL' | awk '{print $3}')"
+            echo "---INFO: Update scripts from repo $Remote_Repo_URL"
+
+            #=======================================
+            mkdir -p ${HOME}/Custler_tmp
+            cp -f ${SCRIPT_DIR}/env.sh ${HOME}/Custler_tmp/
+            cp -f ${SCRIPT_DIR}/TlgChat.json ${HOME}/Custler_tmp/
+            cp -f ${SCRIPT_DIR}/RC_Addr_list.json ${HOME}/Custler_tmp/
+
+            git reset --hard
+            git pull --ff-only
+
+            cp -f  ${HOME}/Custler_tmp/env.sh ${SCRIPT_DIR}/
+            cp -f  ${HOME}/Custler_tmp/TlgChat.json ${SCRIPT_DIR}/
+            cp -f  ${HOME}/Custler_tmp/RC_Addr_list.json ${SCRIPT_DIR}/
+            #=======================================
+
+            cat ${SCRIPT_DIR}/Update_Info.txt
+            echo
+            echo "---INFO: SCRIPTS updated. Files env.sh TlgChat.json RC_Addr_list.json keeped."
+            "${SCRIPT_DIR}/Send_msg_toTelBot.sh" "$HOSTNAME Server" "$Tg_CheckMark $(cat ${SCRIPT_DIR}/Update_Info.txt)" 2>&1 > /dev/null
+            "${SCRIPT_DIR}/Send_msg_toTelBot.sh" "$HOSTNAME Server" "$Tg_CheckMark INFO: SCRIPTS updated. Files env.sh TlgChat.json RC_Addr_list.json keeped." 2>&1 > /dev/null
+        else
+            echo '---WARN: Scripts repo was updated. Please check it and update by hand. If you fully trust me, you can enable autoupdate scripts in env.sh by set variable "Enable_Scripts_Autoupdate" to "true"'
+            "${SCRIPT_DIR}/Send_msg_toTelBot.sh" "$HOSTNAME Server" "$Tg_Warn_sign"+'WARN: Scripts repo was updated. Please check it and update. If you fully trust me, you can enable autoupdate scripts in env.sh by set variable "Enable_Scripts_Autoupdate" to "true"' 2>&1 > /dev/null
+        fi
     fi
 fi
-
 #===========================================================
 # Update NODE
 ${SCRIPT_DIR}/Update_Node_to_new_release.sh
